fix(prefabs): track center eye source for head anchor
The HeadAnchor was originally tracking the Pose Source of Head, but
Unity now warns that if the Pose Source is attached to a camera and
not tracking `Center Eye - HMD Reference` that there may be tracking
issues. So it has been switched over to reduce the possibility of
issues with tracking.